jzftdz.com/requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.1
requests==2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
psutil>=5.9.0
